Remove setters from AppSyncGraphQLRequest so it cant be mutated, and add a newBuilder method
