+++ aws-api/src/main/java/com/amplifyframework/api/aws/AWSApiPlugin.java
@@ -21,6 +21,7 @@ import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 import androidx.core.util.ObjectsCompat;
 
+import com.amplifyframework.AmplifyException;
 import com.amplifyframework.api.ApiException;
 import com.amplifyframework.api.ApiPlugin;
 import com.amplifyframework.api.aws.operation.AWSRestOperation;
@@ -29,6 +30,8 @@ import com.amplifyframework.api.aws.sigv4.DefaultCognitoUserPoolsAuthProvider;
 import com.amplifyframework.api.graphql.GraphQLOperation;
 import com.amplifyframework.api.graphql.GraphQLRequest;
 import com.amplifyframework.api.graphql.GraphQLResponse;
+import com.amplifyframework.api.graphql.Operation;
+import com.amplifyframework.api.graphql.SubscriptionType;
 import com.amplifyframework.api.rest.HttpMethod;
 import com.amplifyframework.api.rest.RestOperation;
 import com.amplifyframework.api.rest.RestOperationRequest;
@@ -36,6 +39,9 @@ import com.amplifyframework.api.rest.RestOptions;
 import com.amplifyframework.api.rest.RestResponse;
 import com.amplifyframework.core.Action;
 import com.amplifyframework.core.Consumer;
+import com.amplifyframework.core.model.AuthRule;
+import com.amplifyframework.core.model.AuthStrategy;
+import com.amplifyframework.core.model.ModelOperation;
 import com.amplifyframework.util.UserAgent;
 
 import org.json.JSONObject;
@@ -43,6 +49,7 @@ import org.json.JSONObject;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
@@ -257,21 +264,27 @@ public final class AWSApiPlugin extends ApiPlugin<Map<String, OkHttpClient>> {
             return null;
         }
 
-        if (graphQLRequest instanceof AppSyncGraphQLRequest) {
+        GraphQLRequest<R> request = graphQLRequest;
+        if (request instanceof AppSyncGraphQLRequest) {
             try {
-                AppSyncGraphQLRequest<R> request = (AppSyncGraphQLRequest<R>) graphQLRequest;
-                if (request.isOwnerArgumentRequired()) {
-                    request.setOwner(getUsername());
+                AppSyncGraphQLRequest<R> appSyncRequest = (AppSyncGraphQLRequest<R>) request;
+                for (AuthRule authRule : appSyncRequest.getModelSchema().getAuthRules()) {
+                    if (isOwnerArgumentRequired(authRule, appSyncRequest.getOperation())) {
+                        request = appSyncRequest.newBuilder()
+                                .variable(authRule.getOwnerFieldOrDefault(), "String!", getUsername())
+                                .build();
+                    }
                 }
-            } catch (ApiException exception) {
-                onSubscriptionFailure.accept(exception);
+            } catch (AmplifyException exception) {
+                onSubscriptionFailure.accept(new ApiException("Failed to set owner field on AppSyncGraphQLRequest",
+                        exception, "See attached exception for details."));
                 return null;
             }
         }
 
         SubscriptionOperation<R> operation = SubscriptionOperation.<R>builder()
             .subscriptionEndpoint(clientDetails.getSubscriptionEndpoint())
-            .graphQlRequest(graphQLRequest)
+            .graphQlRequest(request)
             .responseFactory(gqlResponseFactory)
             .executorService(executorService)
             .onSubscriptionStart(onSubscriptionEstablished)
@@ -281,6 +294,23 @@ public final class AWSApiPlugin extends ApiPlugin<Map<String, OkHttpClient>> {
             .build();
         operation.start();
         return operation;
+    }
+
+    private boolean isOwnerArgumentRequired(AuthRule authRule, Operation operation) {
+        if (!AuthStrategy.OWNER.equals(authRule.getAuthStrategy())) {
+            return false;
+        }
+        List<ModelOperation> operations = authRule.getOperationsOrDefault();
+        if (SubscriptionType.ON_CREATE.equals(operation) && operations.contains(ModelOperation.CREATE)) {
+            return true;
+        }
+        if (SubscriptionType.ON_UPDATE.equals(operation) && operations.contains(ModelOperation.UPDATE)) {
+            return true;
+        }
+        if (SubscriptionType.ON_DELETE.equals(operation) && operations.contains(ModelOperation.DELETE)) {
+            return true;
+        }
+        return false;
     }
 
     private String getUsername() throws ApiException {
