+++ aws-api/src/test/java/com/amplifyframework/api/aws/AWSApiPluginTest.java
@@ -17,17 +17,28 @@ package com.amplifyframework.api.aws;
 
 import androidx.test.core.app.ApplicationProvider;
 
+import com.amplifyframework.AmplifyException;
 import com.amplifyframework.api.ApiException;
 import com.amplifyframework.api.aws.sigv4.CognitoUserPoolsAuthProvider;
+import com.amplifyframework.api.graphql.GraphQLOperation;
 import com.amplifyframework.api.graphql.GraphQLRequest;
 import com.amplifyframework.api.graphql.GraphQLResponse;
+import com.amplifyframework.api.graphql.MutationType;
+import com.amplifyframework.api.graphql.Operation;
 import com.amplifyframework.api.graphql.PaginatedResult;
+import com.amplifyframework.api.graphql.QueryType;
+import com.amplifyframework.api.graphql.SubscriptionType;
 import com.amplifyframework.api.graphql.model.ModelMutation;
 import com.amplifyframework.api.graphql.model.ModelPagination;
 import com.amplifyframework.api.graphql.model.ModelQuery;
 import com.amplifyframework.api.graphql.model.ModelSubscription;
 import com.amplifyframework.core.Consumer;
 import com.amplifyframework.core.NoOpConsumer;
+import com.amplifyframework.core.model.AuthStrategy;
+import com.amplifyframework.core.model.Model;
+import com.amplifyframework.core.model.ModelOperation;
+import com.amplifyframework.core.model.annotations.AuthRule;
+import com.amplifyframework.core.model.annotations.ModelConfig;
 import com.amplifyframework.testmodels.commentsblog.BlogOwner;
 import com.amplifyframework.testmodels.ownerauth.OwnerAuth;
 import com.amplifyframework.testutils.Await;
@@ -58,6 +69,7 @@ import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
@@ -237,24 +249,6 @@ public final class AWSApiPluginTest {
     }
 
     /**
-     * Test that owner variable is added to subscription when needed.
-     * @throws JSONException from JSONAssert.assertEquals
-     */
-    @Test
-    public void graphQlSubscriptionAddsOwnerVariable() throws JSONException {
-        GraphQLRequest<OwnerAuth> request = ModelSubscription.onCreate(OwnerAuth.class);
-        plugin.subscribe(request,
-            NoOpConsumer.create(),
-            NoOpConsumer.create(),
-            NoOpConsumer.create(),
-            EmptyAction.create());
-
-        JSONAssert.assertEquals(Resources.readAsString("request-owner-auth.json"),
-                request.getContent(),
-                true);
-    }
-
-    /**
      * Given that only one API was configured in {@link #setup()},
      * the {@link AWSApiPlugin#getSelectedApiName(EndpointType)} should be able to identify
      * this same API according to just its {@link EndpointType}.
@@ -265,4 +259,126 @@ public final class AWSApiPluginTest {
         String selectedApi = plugin.getSelectedApiName(EndpointType.GRAPHQL);
         assertEquals("graphQlApi", selectedApi);
     }
+
+    /**
+     * Test that request is serialized as expected, with owner variable.
+     * @throws JSONException from JSONAssert.assertEquals
+     */
+    @Test
+    public void ownerArgumentIsAddedAndSerializedInRequest() throws JSONException {
+        GraphQLRequest<OwnerAuth> request = ModelSubscription.onCreate(OwnerAuth.class);
+        GraphQLOperation<OwnerAuth> operation = plugin.subscribe(request,
+                NoOpConsumer.create(),
+                NoOpConsumer.create(),
+                NoOpConsumer.create(),
+                EmptyAction.create());
+
+        JSONAssert.assertEquals(Resources.readAsString("request-owner-auth.json"),
+                operation.getRequest().getContent(),
+                true);
+    }
+
+    /**
+     * Verify that owner argument is required for ON_CREATE subscription if ModelOperation.CREATE is specified.
+     * @throws AmplifyException if a ModelSchema can't be derived from the Model class.
+     */
+    @Test
+    public void ownerArgumentAddedForOnCreate() throws AmplifyException {
+        assertTrue(isOwnerArgumentAdded(Owner.class, SubscriptionType.ON_CREATE));
+        assertTrue(isOwnerArgumentAdded(OwnerCreate.class, SubscriptionType.ON_CREATE));
+    }
+
+    /**
+     * Verify that owner argument is required for ON_UPDATE subscription if ModelOperation.UPDATE is specified.
+     * @throws AmplifyException if a ModelSchema can't be derived from the Model class.
+     */
+    @Test
+    public void ownerArgumentAddedForOnUpdate() throws AmplifyException {
+        assertTrue(isOwnerArgumentAdded(Owner.class, SubscriptionType.ON_UPDATE));
+        assertTrue(isOwnerArgumentAdded(OwnerUpdate.class, SubscriptionType.ON_UPDATE));
+    }
+
+    /**
+     * Verify that owner argument is required for ON_DELETE subscription if ModelOperation.DELETE is specified.
+     * @throws AmplifyException if a ModelSchema can't be derived from the Model class.
+     */
+    @Test
+    public void ownerArgumentAddedForOnDelete() throws AmplifyException {
+        assertTrue(isOwnerArgumentAdded(Owner.class, SubscriptionType.ON_DELETE));
+        assertTrue(isOwnerArgumentAdded(OwnerDelete.class, SubscriptionType.ON_DELETE));
+    }
+
+    /**
+     * Verify owner argument is NOT required if the subscription type is not one of the restricted operations.
+     * @throws AmplifyException if a ModelSchema can't be derived from the Model class.
+     */
+    @Test
+    public void ownerArgumentNotAddedIfOperationNotRestricted() throws AmplifyException {
+        assertFalse(isOwnerArgumentAdded(OwnerCreate.class, SubscriptionType.ON_UPDATE));
+        assertFalse(isOwnerArgumentAdded(OwnerRead.class, SubscriptionType.ON_UPDATE));
+        assertFalse(isOwnerArgumentAdded(OwnerDelete.class, SubscriptionType.ON_UPDATE));
+
+        assertFalse(isOwnerArgumentAdded(OwnerCreate.class, SubscriptionType.ON_DELETE));
+        assertFalse(isOwnerArgumentAdded(OwnerRead.class, SubscriptionType.ON_DELETE));
+        assertFalse(isOwnerArgumentAdded(OwnerUpdate.class, SubscriptionType.ON_DELETE));
+
+        assertFalse(isOwnerArgumentAdded(OwnerRead.class, SubscriptionType.ON_CREATE));
+        assertFalse(isOwnerArgumentAdded(OwnerUpdate.class, SubscriptionType.ON_CREATE));
+        assertFalse(isOwnerArgumentAdded(OwnerDelete.class, SubscriptionType.ON_CREATE));
+    }
+
+    /**
+     * Verify owner argument is NOT added if authStrategy is not OWNER.
+     * @throws AmplifyException if a ModelSchema can't be derived from the Model class.
+     */
+    @Test
+    public void ownerArgumentNotAddedIfNotOwnerStrategy() throws AmplifyException {
+        assertFalse(isOwnerArgumentAdded(Group.class, SubscriptionType.ON_CREATE));
+    }
+
+    /**
+     * Verify owner argument NOT added for Query or Mutation operations.
+     * @throws AmplifyException if a ModelSchema can't be derived from the Model class.
+     */
+    @Test
+    public void verifyOwnerArgumentNotAddedIfNotSubscriptionOperation() throws AmplifyException {
+        assertFalse(isOwnerArgumentAdded(Owner.class, QueryType.GET));
+        assertFalse(isOwnerArgumentAdded(Owner.class, MutationType.CREATE));
+    }
+
+    private boolean isOwnerArgumentAdded(Class<? extends Model> clazz, Operation operation)
+            throws AmplifyException {
+        AppSyncGraphQLRequest<Model> request = AppSyncGraphQLRequest.builder()
+                .modelClass(clazz)
+                .operation(operation)
+                .requestOptions(new ApiGraphQLRequestOptions())
+                .responseType(clazz)
+                .build();
+
+        GraphQLOperation<Model> graphQLOperation = plugin.subscribe(request,
+                NoOpConsumer.create(),
+                NoOpConsumer.create(),
+                NoOpConsumer.create(),
+                EmptyAction.create());
+
+        return "johndoe".equals(graphQLOperation.getRequest().getVariables().get("owner"));
+    }
+
+    @ModelConfig(authRules = { @AuthRule(allow = AuthStrategy.OWNER) })
+    private abstract class Owner implements Model { }
+
+    @ModelConfig(authRules = { @AuthRule(allow = AuthStrategy.OWNER, operations = ModelOperation.CREATE)})
+    private abstract class OwnerCreate implements Model { }
+
+    @ModelConfig(authRules = { @AuthRule(allow = AuthStrategy.OWNER, operations = ModelOperation.READ)})
+    private abstract class OwnerRead implements Model { }
+
+    @ModelConfig(authRules = { @AuthRule(allow = AuthStrategy.OWNER, operations = ModelOperation.UPDATE)})
+    private abstract class OwnerUpdate implements Model { }
+
+    @ModelConfig(authRules = { @AuthRule(allow = AuthStrategy.OWNER, operations = ModelOperation.DELETE)})
+    private abstract class OwnerDelete implements Model { }
+
+    @ModelConfig(authRules = { @AuthRule(allow = AuthStrategy.GROUPS)})
+    private abstract class Group implements Model { }
 }
