--- aws-api/src/test/resources/request-owner-auth-read-update-only.json
@@ -1,4 +1,0 @@
-{
-  "query": "subscription OnUpdateOwnerAuthReadUpdateOnly {onUpdateOwnerAuthReadUpdateOnly {id title}}",
-  "variables": null
-}
